Faster HashMap<MilnorBasisElement, T>

+++ ext/crates/fp/src/prime.rs
@@ -420,7 +420,7 @@ macro_rules! impl_binomial {
 }
 
 impl_binomial!(u32);
-impl_binomial!(u8);
+impl_binomial!(u16);
 impl_binomial!(i32);
 
 pub struct BitflagIterator {
